directory & package refactoring
