+++ backend/tracker-web-app.backend/pom.xml
@@ -2,10 +2,10 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.edu-xored</groupId>
-  <artifactId>tracker-web-app.backend</artifactId>
+  <groupId>edu.xored.tracker</groupId>
+  <artifactId>web-app</artifactId>
   <version>0.1-SNAPSHOT</version>
-  <!--<packaging>jar</packaging>-->
+  <packaging>jar</packaging>
 
   <name>tracker-web-app.backend</name>
   <parent>
@@ -17,8 +17,9 @@
   <dependencies>
     <dependency>
       <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-thymeleaf</artifactId>
+      <artifactId>spring-boot-starter-web</artifactId>
     </dependency>
+
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-devtools</artifactId>
