+++ backend/tracker-web-app.backend/src/main/java/com/edu/xored/tracker/App.java
@@ -1,4 +1,4 @@
-package com.eduxored;
+package com.edu.xored.tracker;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
