+++ backend/tracker-web-app.backend/src/main/java/com/edu/xored/tracker/GreetingsController.java
@@ -1,4 +1,4 @@
-package com.eduxored;
+package com.edu.xored.tracker;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
