+++ backend/tracker-web-app.backend/src/main/resources/static/index.html
@@ -1,10 +1,10 @@
 <!DOCTYPE HTML>
 <html>
 <head>
-    <title>Getting Started: Serving Web Content</title>
+    <title>Simple static page</title>
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
 </head>
 <body>
-<p>I am a static!</p>
+<p>I am static!</p>
 </body>
 </html>
